tools/requirements.txt
# MicroReactor Tools Dependencies
# Install with: pip install -r requirements.txt

pyserial>=3.5
